Misty's Edits
New account page
Add before the "Note" paragraph
<h3>Complete the form below to create a new account with Healthy Kids. Once the account has been created a confirmation and password will be sent to your email address. This may take a few minutes to receive.</h3>
Change text on button from "Create" to "Submit"
Updated Permissions text. I have highlighted the areas that have changed.
You are requesting permission to download the Healthy Kids pediatric obesity risk assessment surveys to your computer. Healthy Kids© is protected by copyright laws and is property of University of California. By downloading Healthy Kids, you agree to keep the original form unchanged. This means that you are not entitled to modify or use the photographs for any other purpose. In addition, all references and logos may not be altered in any way nor removed.You will need to print the survey "as is".
In case of a doubt concerning your planned use of the Healthy Kids surveys or content (photographs, logo or survey questions), please click here to contact us directly.
At the end of the permissions paragraph the "contact us" needs to link to healthykids@ucdavis.edu