Misty's Edits
Educator Survey entry pages (both Healthy Kids & My Child at Meal Time)
Replace "All required fields are in bold and have a red * next to the field label." with:
Enter the first name of the participant. This name with be used to customize the goal sheet. Enter answers for each question. If a participant did not answer a question or checked multiple answers click the "Bypass question" box. When you have finished entering all the questions click the "Submit" button at the bottom of the page. All questions must be answered or "Bypass question" selected.
Remove "Cancel" next to submit button
Remove link "Back to list"