Misty's Edits
Directors Dashboard
Replace the paragraph under "Your Customized Surveys" with this text:
Click on the tool icon to edit the survey you would like tailor. Click the download icon to generate a print quality PDF. You can choose to print now or save the survey file for later printing. The PDF file generated is formatted for professional printing with a .125" bleed on each side.
When clicking the trash can to delete a survey from the table can we add a confirmation pop-up with "Are you sure you want to delete this survey?"